Car Brand entry and value

+++ Models/pricingModel.js
@@ -1,8 +1,7 @@
 module.exports = (sequelize, DataTypes) => {
     const Pricing = sequelize.define("Pricing", {
-      costperhr: { type: DataTypes.INTEGER },
-      carid: { type: DataTypes.INTEGER, primaryKey: true},
+      costperhr: { type: DataTypes.FLOAT },
+      carid: { type: DataTypes.INTEGER, primaryKey: true },
     });
-
     return Pricing;
   };
